Added function to remove html comments
Comments are normally ok, but not needed in ePub files. However,
sometimes you have nested comments, and they are not allowed.
The new function is fairly aggressive in the comment removal, it'll
remove everything from the comment start (<!--) till it reaches the end,
or have found an equal number of coment ends.